+++ src/app/modules/auth/services/rest/user-rest.service.ts
@@ -2,7 +2,7 @@ import {Injectable} from "@angular/core";
 import {HttpClient} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {appLinks} from "../../../../app.links";
-import {User} from "../../../models/user";
+import {User} from "../../../core/models/user";
 import { ChangePasswordForm } from "../../models/change-password-form.model";
 
 @Injectable({
