+++ src/app/modules/cocktails/dish/dish.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import {ModeratorService} from "../../../services/moderator.service";
-import {Dish} from "../../models/dish";
-import {Ingredient} from "../../models/ingredient";
+import {ModeratorService} from "../../core/services/moderator.service";
+import {Dish} from "../../core/models/dish";
+import {Ingredient} from "../../core/models/ingredient";
 import {Cloudinary, CloudinaryImage} from "@cloudinary/url-gen";
 import {thumbnail} from "@cloudinary/url-gen/actions/resize";
 import {byRadius} from "@cloudinary/url-gen/actions/roundCorners";
