+++ src/app/modules/core/models/StockAddDto.ts
