+++ src/app/modules/core/models/friend.ts
