+++ src/app/modules/core/models/stock.model.ts
