+++ src/app/modules/core/services/friend.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import {HttpClient, HttpParams} from "@angular/common/http";
 import {Observable} from "rxjs";
-import {appLinks} from "../app.links";
+import {appLinks} from "../../../app.links";
 
 
 @Injectable({
