+++ src/app/modules/core/services/user.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import {Observable, of} from 'rxjs';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { catchError, map, tap } from 'rxjs/operators';
-import {User} from '../modules/models/user';
+import {User} from '../models/user';
 
 
 @Injectable({
